+++ squiggle_c/squiggle_more.h
@@ -4,7 +4,7 @@
 #define THOUSAND 1000
 #define MILLION (1000 * 1000)
 #define BILLION (1000 * MILLION)
-#define TRILLION (1000 * BILLION)
+#define TRILLION ((uint64_t)1000 * (uint64_t)BILLION)
 
 /* Parallel sampling */
 void sampler_parallel(double (*sampler)(uint64_t* seed), double* results, int n_threads, uint64_t n_samples, int m_seed);
